Before combining person with other entities

+++ Models/Customer.cs
@@ -8,9 +8,9 @@ namespace PharmacyAPI.Models;
 
 public partial class Customer
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
 
-    public long CustomerNumber { get; set; }
+    public int CustomerNumber { get; set; }
 
     public bool IsBlocked { get; set; }
